projet/requirements.txt
# pandas==2.2.2
# scikit-learn==1.2.1
# requests==2.26.0
# lxml==4.6.3
# matplotlib==3.6.3
# seaborn==0.12.2
# numpy==1.23.5
# mlflow==2.3.1

# annotated-types==0.6.0
# anyio==4.3.0
# bcrypt==4.1.2
# certifi==2024.2.2
# cffi==1.16.0
# click==8.1.7
# colorama==0.4.6
# cryptography==42.0.5
# DateTime==5.4
# dnspython==2.6.1
# ecdsa==0.18.0
# email-validator==2.1.0.post1
# exceptiongroup==1.2.0
# fastapi==0.110.0
# h11==0.14.0
# httpcore==1.0.4
# httptools==0.6.1
# httpx==0.27.0
# idna==3.6
# itsdangerous==2.1.2
# Jinja2==3.1.3
# MarkupSafe==2.1.5
# orjson==3.9.15
# passlib==1.7.4
# pyasn1==0.5.1
# pycparser==2.21
# pydantic==2.6.2
# pydantic-extra-types==2.5.0
# pydantic-settings==2.2.1
# pyodbc==5.1.0
# python-dotenv==1.0.1
# python-jose==3.3.0
# python-multipart==0.0.9
# # pytz==2024.1
# PyYAML==6.0.1
# rsa==4.9
# six==1.16.0
# sniffio==1.3.1
# starlette==0.36.3
# typing_extensions==4.10.0
# ujson==5.9.0
# uvicorn==0.27.1
# watchfiles==0.21.0
# websockets==12.0
# zope.interface==6.2



alembic==1.13.1
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker==7.0.0
ecdsa==0.19.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
Flask==3.0.3
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.12.1
nest-asyncio==1.6.0
numpy==1.23.5
oauthlib==3.2.2
orjson==3.9.15
packaging==23.2
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.9.0b1
Pygments==2.17.2
PyJWT==2.8.0
pyodbc==5.1.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.4
# pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
querystring-parser==1.2.4
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.29.0
waitress==2.1.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
zipp==3.18.1
zope.interface==6.2